Refactor match service and controller
- Update MatchService interface to include a method for getting the current match for a user
- Remove unused Matchmaking service and implementation files
- Update MatchesController to import MatchService from the correct package
- Update MatchServiceImpl to include logic for getting the current match for a user
- Add PlayerProfileRepository dependency to MatchServiceImpl
