+++ backend/src/main/java/com/project/G1_T3/match/service/MatchService.java
@@ -5,6 +5,7 @@ import com.project.G1_T3.match.model.MatchDTO;
 
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import java.util.UUID;
 
 public interface MatchService {
 
@@ -13,5 +14,7 @@ public interface MatchService {
     public void startMatch(@PathVariable Long matchId, @RequestBody MatchDTO matchDTO);
 
     public void completeMatch(@PathVariable Long matchId, @RequestBody MatchDTO matchDTO);
+
+    public Match getCurrentMatchForUser(UUID userId);
 
 }
