+++ backend/src/main/java/com/project/G1_T3/matchmaking/controller/MatchesController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.project.G1_T3.match.model.Match;
-import com.project.G1_T3.matchmaking.service.MatchService;
+import com.project.G1_T3.match.service.MatchService;
 
 @RestController
 @RequestMapping("/matches")
